requirements.txt
Django==3.2.12
django-admin-sortable2==1.0.0

